[BACKEND] Split up TritonGPUToLLVM.cpp; renamed TritonGPUToLLVMPass.cpp -> TritonGPUToLLVM.cpp (#3055)
