+++ lib/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -13,8 +13,13 @@ add_triton_library(TritonGPUToLLVM
     LoadStoreOpToLLVM.cpp
     BarrierOpToLLVM.cpp
     TritonGPUToLLVM.cpp
-    TritonGPUToLLVMPass.cpp
     DecomposeUnsupportedConversions.cpp
+    AssertOpToLLVM.cpp
+    PrintOpToLLVM.cpp
+    MemoryOpToLLVM.cpp
+    ControlFlowOpToLLVM.cpp
+    MakeRangeOpToLLVM.cpp
+    SPMDOpToLLVM.cpp
     ReduceOpToLLVM.cpp
     ScanOpToLLVM.cpp
     TypeConverter.cpp
