+++ lib/Conversion/TritonGPUToLLVM/ClusterOpsToLLVM.cpp
@@ -59,8 +59,8 @@ struct ClusterWaitOpConversion
 } // namespace
 
 void mlir::triton::populateClusterOpsToLLVMPatterns(
-    LLVMTypeConverter &typeConverter, RewritePatternSet &patterns, int numWarps,
-    ModuleAxisInfoAnalysis &axisInfoAnalysis, PatternBenefit benefit) {
+    LLVMTypeConverter &typeConverter, RewritePatternSet &patterns,
+    PatternBenefit benefit) {
   patterns.add<ClusterArriveOpConversion>(typeConverter, benefit);
   patterns.add<ClusterWaitOpConversion>(typeConverter, benefit);
   return;
