+++ lib/Conversion/TritonGPUToLLVM/MakeRangeOpToLLVM.cpp
@@ -1,0 +1,50 @@
+#include "PatternTritonGPUOpToLLVM.h"
+#include "Utility.h"
+
+namespace {
+
+using namespace mlir;
+using namespace mlir::triton;
+
+struct MakeRangeOpConversion
+    : public ConvertOpToLLVMPattern<triton::MakeRangeOp> {
+
+  MakeRangeOpConversion(LLVMTypeConverter &converter, PatternBenefit benefit)
+      : ConvertOpToLLVMPattern<triton::MakeRangeOp>(converter, benefit) {}
+
+  LogicalResult
+  matchAndRewrite(triton::MakeRangeOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op->getLoc();
+    auto rankedTy = op.getResult().getType().cast<RankedTensorType>();
+    auto shape = rankedTy.getShape();
+    auto layout = rankedTy.getEncoding();
+
+    auto elemTy = rankedTy.getElementType();
+    assert(elemTy.isInteger(32));
+    Value start = createIndexAttrConstant(rewriter, loc, elemTy, op.getStart());
+    auto idxs = emitIndices(loc, rewriter, layout, rankedTy, true);
+    unsigned elems = idxs.size();
+    SmallVector<Value> retVals(elems);
+    // TODO: slice layout has more elements than expected.
+    // Unexpected behavior for make range, but generally OK when followed by
+    // expand dims + broadcast. very weird behavior otherwise potentially.
+    for (const auto &multiDim : llvm::enumerate(idxs)) {
+      assert(multiDim.value().size() == 1);
+      retVals[multiDim.index()] = add(multiDim.value()[0], start);
+    }
+    auto typeConverter = getTypeConverter();
+    Value result =
+        packLLElements(loc, typeConverter, retVals, rewriter, rankedTy);
+    rewriter.replaceOp(op, result);
+    return success();
+  }
+};
+
+} // namespace
+
+void mlir::triton::populateMakeRangeOpToLLVMPattern(
+    LLVMTypeConverter &typeConverter, RewritePatternSet &patterns,
+    PatternBenefit benefit) {
+  patterns.add<MakeRangeOpConversion>(typeConverter, benefit);
+}
