+++ lib/Conversion/TritonGPUToLLVM/MemoryOpToLLVM.cpp
@@ -1,0 +1,69 @@
+#include "PatternTritonGPUOpToLLVM.h"
+#include "Utility.h"
+
+namespace {
+
+using namespace mlir;
+using namespace mlir::triton;
+
+struct AllocTensorOpConversion
+    : public ConvertOpToLLVMPattern<triton::gpu::AllocTensorOp> {
+  using ConvertOpToLLVMPattern<
+      triton::gpu::AllocTensorOp>::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(triton::gpu::AllocTensorOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op->getLoc();
+    Value smemBase =
+        LLVM::getSharedMemoryBase(loc, rewriter, op.getOperation());
+    auto resultTy = op.getType().dyn_cast<RankedTensorType>();
+    auto elemPtrTy = ptr_ty(rewriter.getContext(), 3);
+    auto typeConverter = getTypeConverter();
+    smemBase = bitcast(smemBase, elemPtrTy);
+    auto sharedLayout =
+        resultTy.getEncoding().cast<triton::gpu::SharedEncodingAttr>();
+    auto order = sharedLayout.getOrder();
+    // Workaround for 3D tensors
+    // TODO: we need to modify the pipeline pass to give a proper shared
+    // encoding to 3D tensors
+    SmallVector<unsigned> newOrder;
+    if (resultTy.getShape().size() != order.size()) {
+      for (auto i = 0; i < order.size(); ++i)
+        newOrder.push_back(order[i] + 1);
+      newOrder.push_back(0);
+    } else {
+      newOrder = SmallVector<unsigned>(order.begin(), order.end());
+    }
+
+    auto llvmElemTy = typeConverter->convertType(resultTy.getElementType());
+    auto shapePerCTA = getShapePerCTA(sharedLayout, resultTy.getShape());
+    auto smemObj = SharedMemoryObject(smemBase, llvmElemTy, shapePerCTA,
+                                      newOrder, loc, rewriter);
+    auto retVal = getStructFromSharedMemoryObject(loc, smemObj, rewriter);
+    rewriter.replaceOp(op, retVal);
+    return success();
+  }
+};
+
+struct DeallocTensorOpConversion
+    : public ConvertOpToLLVMPattern<triton::gpu::DeallocTensorOp> {
+  using ConvertOpToLLVMPattern<
+      triton::gpu::DeallocTensorOp>::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(triton::gpu::DeallocTensorOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+} // namespace
+
+void mlir::triton::populateMemoryOpToLLVMPattern(
+    LLVMTypeConverter &typeConverter, RewritePatternSet &patterns,
+    PatternBenefit benefit) {
+  patterns.add<AllocTensorOpConversion>(typeConverter, benefit);
+  patterns.add<DeallocTensorOpConversion>(typeConverter, benefit);
+}
