+++ lib/Conversion/TritonGPUToLLVM/TritonGPUToLLVM.cpp
@@ -1,873 +1,401 @@
-#include "PatternTritonGPUOpToLLVM.h"
-#include "Utility.h"
+#include "mlir/Analysis/DataFlowFramework.h"
+#include "mlir/Conversion/ArithToLLVM/ArithToLLVM.h"
+#include "mlir/Conversion/ControlFlowToLLVM/ControlFlowToLLVM.h"
+#include "mlir/Conversion/GPUToNVVM/GPUToNVVMPass.h"
+#include "mlir/Conversion/LLVMCommon/VectorPattern.h"
+#include "mlir/Conversion/MathToLLVM/MathToLLVM.h"
+#include "mlir/Conversion/SCFToControlFlow/SCFToControlFlow.h"
 #include "mlir/Dialect/ControlFlow/IR/ControlFlowOps.h"
+#include "mlir/Dialect/Index/IR/IndexDialect.h"
+#include "mlir/Dialect/Index/IR/IndexOps.h"
+#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
+#include "mlir/Dialect/LLVMIR/NVVMDialect.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+#include "triton/Analysis/Allocation.h"
+#include "triton/Analysis/AxisInfo.h"
+#include "triton/Analysis/Membar.h"
+#include "triton/Conversion/TritonGPUToLLVM/Passes.h"
+#include "triton/Dialect/NVGPU/IR/Dialect.h"
+#include "triton/Dialect/Triton/IR/Dialect.h"
+#include "triton/Dialect/TritonGPU/IR/Dialect.h"
+#include "triton/Dialect/TritonNvidiaGPU/IR/Dialect.h"
+#include "triton/Tools/Sys/GetPlatform.hpp"
 
-namespace {
+#include "PatternTritonGPUOpToLLVM.h"
+#include "TypeConverter.h"
+#include "Utility.h"
+
+namespace mlir {
+namespace triton {
+#define GEN_PASS_DEF_CONVERTTRITONGPUTOLLVM
+#include "triton/Conversion/TritonGPUToLLVM/Passes.h.inc"
+} // namespace triton
+} // namespace mlir
+
+namespace mlir {
+FailureOr<LLVM::LLVMFuncOp>
+convertFuncOpToLLVMFuncOp(FunctionOpInterface funcOp,
+                          ConversionPatternRewriter &rewriter,
+                          const LLVMTypeConverter &converter);
+}
 
 using namespace mlir;
 using namespace mlir::triton;
+namespace ttng = mlir::triton::nvidia_gpu;
 
-using ::mlir::LLVM::getSharedMemoryObjectFromStruct;
-using ::mlir::LLVM::getSRegValue;
-using ::mlir::triton::gpu::getTotalElemsPerThread;
-using ::mlir::triton::gpu::SharedEncodingAttr;
+namespace {
 
-Value llGetPid(int axis, Location loc, ModuleOp moduleOp,
-               ConversionPatternRewriter &rewriter) {
-  assert(axis >= 0);
-  assert(axis < 3);
-  assert(moduleOp);
-
-  // It is not easy to get the compute capability here, so we use numCTAs to
-  // decide the semantic of GetProgramIdOp. If numCTAs = 1, then
-  // GetProgramIdOp is converted to "%ctaid", otherwise it is converted to
-  // "%clusterid".
-  int numCTAs = triton::gpu::TritonGPUDialect::getNumCTAs(moduleOp);
-
-  std::string sreg = numCTAs == 1 ? "%ctaid." : "%clusterid.";
-  sreg.append(1, 'x' + axis); // 0 -> 'x', 1 -> 'y', 2 -> 'z'
-  return getSRegValue(rewriter, loc, sreg);
+// pass ws related named attrs.
+static void addAttrs(Operation *op, ArrayRef<mlir::NamedAttribute> attrs) {
+  for (const NamedAttribute attr : attrs)
+    op->setAttr(attr.getName(), attr.getValue());
 }
 
-struct ReturnOpConversion : public ConvertOpToLLVMPattern<triton::ReturnOp> {
-  using ConvertOpToLLVMPattern<triton::ReturnOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::ReturnOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    unsigned numArguments = op.getNumOperands();
-
-    // Currently, Triton kernel function always return nothing.
-    // TODO(Superjomn) add support for non-inline device function
-    if (numArguments > 0) {
-      return rewriter.notifyMatchFailure(
-          op, "Only kernel function with nothing returned is supported.");
-    }
-
-    rewriter.replaceOpWithNewOp<LLVM::ReturnOp>(op, TypeRange(), ValueRange(),
-                                                op->getAttrs());
-    return success();
+class TritonLLVMFunctionConversionTarget : public ConversionTarget {
+public:
+  explicit TritonLLVMFunctionConversionTarget(MLIRContext &ctx, Target target)
+      : ConversionTarget(ctx) {
+    addLegalDialect<index::IndexDialect>();
+    addLegalDialect<LLVM::LLVMDialect>();
+    addLegalDialect<NVVM::NVVMDialect>();
+    addLegalOp<mlir::UnrealizedConversionCastOp>();
   }
 };
 
-struct BroadcastOpConversion
-    : public ConvertOpToLLVMPattern<triton::BroadcastOp> {
-  using ConvertOpToLLVMPattern<triton::BroadcastOp>::ConvertOpToLLVMPattern;
+class FoldSplatMaskInInsertAsync : public mlir::RewritePattern {
+
+public:
+  FoldSplatMaskInInsertAsync(mlir::MLIRContext *context)
+      : mlir::RewritePattern(
+            triton::nvidia_gpu::InsertSliceTMAOp::getOperationName(), 1,
+            context) {}
 
   LogicalResult
-  matchAndRewrite(triton::BroadcastOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    // Following the order of indices in the legacy code, a broadcast of:
-    //   [s(0), s(1) ... s(k-1),    1, s(k+1), s(k+2) ... s(n-1)]
-    // =>
-    //   [s(0), s(1) ... s(k-1), s(k), s(k+1), s(k+2) ... s(n-1)]
-    //
-    // logically maps to a broadcast within a thread's scope:
-    //   [cta(0)..cta(k-1),     1,cta(k+1)..cta(n-1),spt(0)..spt(k-1),
-    //   1,spt(k+1)..spt(n-1)]
-    // =>
-    //   [cta(0)..cta(k-1),cta(k),cta(k+1)..cta(n-1),spt(0)..spt(k-1),spt(k),spt(k+1)..spt(n-1)]
-    //
-    // regardless of the order of the layout
-    //
-    Location loc = op->getLoc();
-    Value src = adaptor.getSrc();
-    Value result = op.getResult();
-    auto srcTy = op.getSrc().getType().cast<RankedTensorType>();
-    auto resultTy = result.getType().cast<RankedTensorType>();
-    auto srcLayout = srcTy.getEncoding();
-    auto resultLayout = resultTy.getEncoding();
-    auto srcShape = srcTy.getShape();
-    auto resultShape = resultTy.getShape();
-    unsigned rank = srcTy.getRank();
-    auto typeConverter = getTypeConverter();
-
-    assert(rank == resultTy.getRank());
-    auto order = triton::gpu::getOrder(srcLayout);
-    auto srcOffsets = emitOffsetForLayout(srcLayout, srcTy);
-    auto resultOffsets = emitOffsetForLayout(resultLayout, resultTy);
-    SmallVector<Value> srcVals = unpackLLElements(loc, src, rewriter);
-
-    std::map<SmallVector<unsigned>, Value> srcValues;
-    for (size_t i = 0; i < srcOffsets.size(); i++) {
-      srcValues[srcOffsets[i]] = srcVals[i];
-    }
-
-    SmallVector<Value> resultVals;
-    for (size_t i = 0; i < resultOffsets.size(); i++) {
-      auto offset = resultOffsets[i];
-      for (size_t j = 0; j < srcShape.size(); j++)
-        if (srcShape[j] == 1)
-          offset[j] = 0;
-      resultVals.push_back(srcValues.at(offset));
-    }
-
-    Value resultStruct =
-        packLLElements(loc, typeConverter, resultVals, rewriter, resultTy);
-    rewriter.replaceOp(op, {resultStruct});
-    return success();
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override {
+    auto insertOp = cast<triton::nvidia_gpu::InsertSliceTMAOp>(op);
+    if (!insertOp.getMask())
+      return failure();
+    auto splatOp = insertOp.getMask().getDefiningOp<triton::SplatOp>();
+    if (!splatOp)
+      return failure();
+    rewriter.modifyOpInPlace(insertOp, [&]() {
+      insertOp.getMaskMutable().assign(splatOp->getOperand(0));
+    });
+    return mlir::success();
   }
 };
 
-// The input print op contains:
-//  - a "prefix" (string) specified by the user, and
-//  - one or more "operands" (tensors).
-//
-// For each operand, we print all of the values contained in this GPU thread,
-// one per line, along with the index of the value in its tensor.
-struct PrintOpConversion : public ConvertOpToLLVMPattern<triton::PrintOp> {
-  using ConvertOpToLLVMPattern<triton::PrintOp>::ConvertOpToLLVMPattern;
+/// FuncOp legalization pattern that converts MemRef arguments to pointers to
+/// MemRef descriptors (LLVM struct data types) containing all the MemRef type
+/// information.
+
+struct FuncOpConversion : public ConvertOpToLLVMPattern<triton::FuncOp> {
+  FuncOpConversion(LLVMTypeConverter &converter, int numWarps,
+                   PatternBenefit benefit)
+      : ConvertOpToLLVMPattern(converter, benefit), numWarps(numWarps) {}
+
+  /// Only retain those attributes that are not constructed by
+  /// `LLVMFuncOp::build`. If `filterArgAttrs` is set, also filter out argument
+  /// attributes.
+  static void filterFuncAttributes(triton::FuncOp op, bool filterArgAttrs,
+                                   SmallVectorImpl<NamedAttribute> &result) {
+
+    for (const auto &attr : op->getAttrs()) {
+      if (attr.getName() == SymbolTable::getSymbolAttrName() ||
+          attr.getName() == op.getFunctionTypeAttrName() ||
+          attr.getName() == "std.varargs" ||
+          (filterArgAttrs && attr.getName() == op.getArgAttrsAttrName()))
+        continue;
+      result.push_back(attr);
+    }
+  }
+
+  triton::FuncOp amendFuncOp(triton::FuncOp funcOp,
+                             ConversionPatternRewriter &rewriter) const {
+    // Push back a variable that indicates the current stack pointer of shared
+    // memory to the function arguments.
+    auto loc = funcOp.getLoc();
+    auto ctx = funcOp->getContext();
+    auto ptrTy = LLVM::LLVMPointerType::get(rewriter.getContext(), 3);
+    // 1. Modify the function type to add the new argument.
+    auto funcTy = funcOp.getFunctionType();
+    auto amendedInputTy = llvm::to_vector<4>(funcTy.getInputs());
+    amendedInputTy.push_back(ptrTy);
+    auto amendedFuncTy = FunctionType::get(funcTy.getContext(), amendedInputTy,
+                                           funcTy.getResults());
+    // 2. Modify the argument attributes to add the new argument.
+    SmallVector<NamedAttribute> amendedAttrs;
+    filterFuncAttributes(funcOp, /*filterArgAttrs=*/true, amendedAttrs);
+    auto amendedArgAttrs = llvm::to_vector<4>(funcOp.getAllArgAttrs());
+    amendedArgAttrs.emplace_back(DictionaryAttr::get(ctx));
+    amendedAttrs.push_back(rewriter.getNamedAttr(
+        funcOp.getArgAttrsAttrName(), rewriter.getArrayAttr(amendedArgAttrs)));
+    // 3. Add a new argument to the region
+    auto amendedFuncOp = rewriter.create<triton::FuncOp>(
+        funcOp.getLoc(), funcOp.getName(), amendedFuncTy, amendedAttrs);
+    auto &region = funcOp.getBody();
+    region.addArgument(ptrTy, loc);
+    rewriter.inlineRegionBefore(region, amendedFuncOp.getBody(),
+                                amendedFuncOp.end());
+    return amendedFuncOp;
+  }
 
   LogicalResult
-  matchAndRewrite(triton::PrintOp op, OpAdaptor adaptor,
+  matchAndRewrite(triton::FuncOp funcOp, OpAdaptor adaptor,
                   ConversionPatternRewriter &rewriter) const override {
-    auto typeConverter = getTypeConverter();
-    auto loc = op->getLoc();
-    Value prefixStr =
-        LLVM::addStringToModule(loc, rewriter, "printfPrefix_", op.getPrefix());
+    // Prevent LLVM's inliner to inline this function
+    auto amendedFuncOp = funcOp;
+    if (!LLVM::isKernel(funcOp))
+      amendedFuncOp = amendFuncOp(funcOp, rewriter);
 
-    auto getPid = [&](int axis) {
-      return llGetPid(axis, loc, op->getParentOfType<ModuleOp>(), rewriter);
-    };
-    std::array<Value, 3> pid = {getPid(0), getPid(1), getPid(2)};
+    LLVM::LLVMFuncOp newFuncOp = *mlir::convertFuncOpToLLVMFuncOp(
+        amendedFuncOp, rewriter, *getTypeConverter());
+    if (!newFuncOp) {
+      return failure();
+    }
 
-    // Simple printf of a string without any tensors.
-    if (op.getNumOperands() == 0) {
-      std::string formatStr;
-      llvm::raw_string_ostream os(formatStr);
-      os << "pid (" << getFormatSubstr(pid[0]) << ", "
-         << getFormatSubstr(pid[1]) << ", " << getFormatSubstr(pid[2]) << ")%s";
-      llPrintf(formatStr, {pid[0], pid[1], pid[2], prefixStr}, rewriter);
+    auto ctx = funcOp->getContext();
+
+    if (LLVM::isKernel(funcOp)) {
+      // Set an attribute to indicate this function is a kernel entry.
+      newFuncOp->setAttr("nvvm.kernel",
+                         rewriter.getIntegerAttr(type::u1Ty(ctx), 1));
     } else {
-      for (size_t i = 0; i < op.getNumOperands(); i++) {
-        // Elements of the tensor that are resident in this GPU thread.
-        auto elems = unpackLLElements(loc, adaptor.getOperands()[i], rewriter);
+      // The noinline attribute will be used by the LLVM codegen to prevent
+      // inlining.
+      // https://github.com/llvm/llvm-project/blob/main/mlir/lib/Dialect/LLVMIR/IR/LLVMInlining.cpp#L267
+      newFuncOp.setPassthroughAttr(
+          ArrayAttr::get(ctx, rewriter.getStringAttr("noinline")));
+      rewriter.eraseOp(amendedFuncOp);
+    }
+    // Set an attribute for maxntidx, it could be used in latter LLVM codegen
+    // for `nvvm.annotation` metadata.
+    newFuncOp->setAttr("nvvm.maxntid",
+                       rewriter.getDenseI32ArrayAttr(32 * numWarps));
 
-        // Get the indices of `elems` within the tensor.  Note that if `elems`
-        // has an "interesting" layout, then these will not be in any
-        // particularly nice order.
+    // required by AxisInfoAnalysis
+    rewriter.eraseOp(funcOp);
+    return success();
+  }
 
-        // Extract the shape of the tensor being printed and use it to figure
-        // out how many digits we need for each of the dimensions.
-        SmallVector<int, 8> dimWidths;
-        SmallVector<SmallVector<Value>> indices;
-        if (auto rankedTy =
-                op.getOperand(i).getType().dyn_cast<RankedTensorType>()) {
-          indices = emitIndices(loc, rewriter, rankedTy.getEncoding(), rankedTy,
-                                true);
-          for (int64_t dim : rankedTy.getShape()) {
-            if (dim > 0) {
-              dimWidths.push_back(static_cast<int>(std::ceil(std::log10(dim))));
-            } else {
-              dimWidths.push_back(0);
-            }
+private:
+  int numWarps{0};
+};
+
+class TritonLLVMConversionTarget : public ConversionTarget {
+public:
+  explicit TritonLLVMConversionTarget(MLIRContext &ctx, Target target)
+      : ConversionTarget(ctx) {
+    addLegalDialect<LLVM::LLVMDialect>();
+    addLegalDialect<NVVM::NVVMDialect>();
+    addLegalDialect<mlir::triton::nvgpu::NVGPUDialect>();
+    addIllegalDialect<triton::TritonDialect>();
+    addIllegalDialect<triton::gpu::TritonGPUDialect>();
+    addIllegalDialect<triton::nvidia_gpu::TritonNvidiaGPUDialect>();
+    addIllegalDialect<mlir::gpu::GPUDialect>();
+    addLegalOp<mlir::UnrealizedConversionCastOp>();
+  }
+};
+
+struct ConvertTritonGPUToLLVM
+    : public triton::impl::ConvertTritonGPUToLLVMBase<ConvertTritonGPUToLLVM> {
+  using ConvertTritonGPUToLLVMBase::ConvertTritonGPUToLLVMBase;
+
+  void getDependentDialects(DialectRegistry &registry) const override {
+    registry.insert<triton::nvgpu::NVGPUDialect, LLVM::LLVMDialect,
+                    NVVM::NVVMDialect>();
+  }
+
+  ConvertTritonGPUToLLVM(int32_t computeCapability, Target target,
+                         mlir::triton::gpu::TMAMetadataTy *tmaMetadata)
+      : ConvertTritonGPUToLLVMBase({computeCapability, target}),
+        tmaMetadata(tmaMetadata) {}
+
+  void runOnOperation() override {
+    MLIRContext *context = &getContext();
+    ModuleOp mod = getOperation();
+
+    mlir::LowerToLLVMOptions option(context);
+    option.overrideIndexBitwidth(32);
+    TritonGPUToLLVMTypeConverter typeConverter(context, option);
+    TritonLLVMConversionTarget convTarget(*context, target);
+    int numWarps = triton::gpu::TritonGPUDialect::getNumWarps(mod);
+    int numCTAs = triton::gpu::TritonGPUDialect::getNumCTAs(mod);
+    int threadsPerWarp = triton::gpu::TritonGPUDialect::getThreadsPerWarp(mod);
+
+    // Hack: WSMaterialization may have changed the effective number of warps,
+    // in a way that isn't reflected in triton_gpu.num-warps.  If so, we have to
+    // respect that here.
+    if (Attribute attr = mod->getAttr("triton_gpu.num-warp-groups-per-cta")) {
+      numWarps *= attr.cast<IntegerAttr>().getInt();
+    }
+
+    // Allocate shared memory and set barrier
+    ModuleAllocation allocation(mod);
+    ModuleMembarAnalysis membarPass(&allocation);
+    membarPass.run();
+
+    /* Get tensorPtrMap before conversion */
+    TensorPtrMapT tensorPtrMap;
+    mod.walk(
+        [&tensorPtrMap](mlir::triton::nvidia_gpu::InsertSliceTMAOp insertOp) {
+          auto src = insertOp.getSrc();
+          auto ptrTy = src.getType().dyn_cast<triton::PointerType>();
+          if (ptrTy && ptrTy.getPointeeType().isa<RankedTensorType>()) {
+            auto makeTensorPtrOp = getMakeTensorPtrOp(insertOp.getSrc());
+            tensorPtrMap[insertOp.getOperation()] = makeTensorPtrOp;
           }
-        } else {
-          // We're printing a scalar.
-          assert(elems.size() == 1);
-          indices.push_back({});
-        }
+        });
 
-        if (!elems.empty()) {
-          printTensor(prefixStr, /*operand=*/i,
-                      /*numOperands=*/op.getNumOperands(), elems, pid, indices,
-                      dimWidths, rewriter);
-        }
-      }
+    mod.walk(
+        [&tensorPtrMap](mlir::triton::nvidia_gpu::StoreAsyncTMAOp storeOp) {
+          auto dst = storeOp.getDst();
+          auto ptrTy = dst.getType().dyn_cast<triton::PointerType>();
+          if (ptrTy && ptrTy.getPointeeType().isa<RankedTensorType>()) {
+            auto makeTensorPtrOp = getMakeTensorPtrOp(storeOp.getDst());
+            tensorPtrMap[storeOp.getOperation()] = makeTensorPtrOp;
+          }
+        });
+
+    // Hack: cleanup
+    {
+      RewritePatternSet patterns(context);
+      patterns.add<FoldSplatMaskInInsertAsync>(context);
+      SmallVector<Operation *> insertSlices;
+      mod.walk([&insertSlices](triton::nvidia_gpu::InsertSliceTMAOp op) {
+        insertSlices.push_back(op);
+      });
+      if (applyOpPatternsAndFold(insertSlices, std::move(patterns)).failed())
+        signalPassFailure();
     }
-    rewriter.eraseOp(op);
-    return success();
+
+    // Lower functions
+    {
+      mlir::LowerToLLVMOptions option(context);
+      TritonGPUToLLVMTypeConverter typeConverter(context, option);
+      TritonLLVMFunctionConversionTarget funcTarget(*context, target);
+      RewritePatternSet funcPatterns(context);
+      funcPatterns.add<FuncOpConversion>(typeConverter, numWarps,
+                                         /*benefit=*/1);
+      mlir::cf::populateControlFlowToLLVMConversionPatterns(typeConverter,
+                                                            funcPatterns);
+      if (failed(
+              applyPartialConversion(mod, funcTarget, std::move(funcPatterns))))
+        return signalPassFailure();
+    }
+
+    // initSharedMemory is run before the conversion of call and ret ops,
+    // because the call op has to know the shared memory base address of each
+    // function
+    initSharedMemory(typeConverter);
+    ModuleAxisInfoAnalysis axisInfoAnalysis(mod);
+
+    // Emit logics to get threadId/blockIds/linearized clusterCTAId etc. and
+    // cache the values. The reason to do it here is that cluster_ctaid is
+    // currently implemented via inline asm, and thus cannot be CSEed.
+    // clusterCTAId will be emitted only when numCTAs is larger than 1, and
+    // other values will be DCEed if not used hereafter.
+    bool isWarpSpecialization =
+        ttng::TritonNvidiaGPUDialect::getWSSupportedAttr(mod);
+    OpBuilder::InsertPoint indexInsertPoint;
+
+    // tmaMetadata is absent in a triton-opt unit test, in this case, create a
+    // local one and dump it after this pass is done.
+    mlir::triton::gpu::TMAMetadataTy tmaMetaDataDebug;
+    if (tmaMetadata == nullptr)
+      tmaMetadata = &tmaMetaDataDebug;
+
+    RewritePatternSet patterns(context);
+    int benefit = 10;
+    populateConvertLayoutOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populateDotOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populateElementwiseOpToLLVMPatterns(
+        typeConverter, patterns, axisInfoAnalysis, computeCapability, benefit);
+    populateLoadStoreOpToLLVMPatterns(typeConverter, patterns, axisInfoAnalysis,
+                                      tmaMetadata, &tensorPtrMap, benefit);
+    populateReduceOpToLLVMPatterns(typeConverter, patterns, computeCapability,
+                                   benefit);
+    populateScanOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populateViewOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populateBarrierOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populateTensorPtrOpsToLLVMPatterns(typeConverter, patterns, benefit);
+    populateClusterOpsToLLVMPatterns(typeConverter, patterns, benefit);
+    populateRegReallocOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populateHistogramOpToLLVMPatterns(typeConverter, patterns, benefit);
+    populatePrintOpToLLVMPattern(typeConverter, patterns, benefit);
+    populateAssertOpToLLVMPattern(typeConverter, patterns, benefit);
+    populateMemoryOpToLLVMPattern(typeConverter, patterns, benefit);
+    populateControlFlowOpToLLVMPattern(typeConverter, patterns, benefit);
+    populateMakeRangeOpToLLVMPattern(typeConverter, patterns, benefit);
+    populateSPMDOpToLLVMPattern(typeConverter, patterns, benefit);
+    // TODO(thomas): this should probably be done in a separate step to not
+    // interfere with our own lowering of arith ops. Add arith/math's patterns
+    // to help convert scalar expression to LLVM.
+    mlir::arith::populateArithToLLVMConversionPatterns(typeConverter, patterns);
+    mlir::populateMathToLLVMConversionPatterns(typeConverter, patterns);
+    mlir::populateGpuToNVVMConversionPatterns(typeConverter, patterns);
+    mlir::cf::populateControlFlowToLLVMConversionPatterns(typeConverter,
+                                                          patterns);
+    if (failed(applyPartialConversion(mod, convTarget, std::move(patterns))))
+      return signalPassFailure();
+
+    // Fold CTAId when there is only 1 CTA.
+    if (numCTAs == 1) {
+      mod.walk([](triton::nvgpu::ClusterCTAIdOp id) {
+        OpBuilder b(id);
+        Value zero = LLVM::createConstantI32(id->getLoc(), b, 0);
+        id.replaceAllUsesWith(zero);
+      });
+    }
   }
 
-  void printTensor(Value prefixStr, size_t operand, size_t numOperands,
-                   ArrayRef<Value> elems, std::array<Value, 3> pid,
-                   ArrayRef<SmallVector<Value>> indices,
-                   ArrayRef<int> dimWidths,
-                   ConversionPatternRewriter &rewriter) const {
-    assert(!elems.empty());
-    assert(elems.size() == indices.size());
-    assert(dimWidths.size() == indices.front().size());
+private:
+  mlir::triton::gpu::TMAMetadataTy *tmaMetadata = nullptr;
 
-    size_t rank = dimWidths.size();
-
-    // Format is:
-    //   pid (<x>, <y>, <z>) idx (<i1>, <i2>, ...)<prefix> (operand <n>) <elem>
-    // where we leave off "(operand <n>)" if there's only one operand.
+  void initSharedMemory(LLVMTypeConverter &typeConverter) {
+    ModuleOp mod = getOperation();
+    OpBuilder b(mod.getBodyRegion());
+    auto ctx = mod.getContext();
+    auto loc = mod.getLoc();
+    auto elemTy = typeConverter.convertType(b.getIntegerType(8));
+    // Set array size 0 and external linkage indicates that we use dynamic
+    // shared allocation to allow a larger shared memory size for each kernel.
     //
-    // The Python wrapper munges `prefix` so that it prints nicely (e.g. starts
-    // with " " and ends with ": ").
-
-    Value formatStrValue;
-    for (int i = 0; i < elems.size(); i++) {
-      std::string formatStr;
-      llvm::raw_string_ostream os(formatStr);
-
-      // nvptx printf can only accept 32 args; if we pass more than that, it
-      // will print garbage for the trailing args.
-      constexpr int kMaxPrintfOperands = 32;
-      SmallVector<Value, kMaxPrintfOperands> printfOperands;
-
-      // TODO(jlebar): We really should pad the pid, but because the max pid is
-      // not known at compile-time, this would require nontrivial device-side
-      // work.
-      os << "pid (";
-      for (int j = 0; j < pid.size(); j++) {
-        if (j != 0) {
-          os << ", ";
-        }
-        os << getFormatSubstr(pid[j]);
-        printfOperands.push_back(pid[j]);
-      }
-      os << ") ";
-
-      // If `rank` is large enough, we could end up exceeding
-      // kMaxPrintfOperands.  In that case, just truncate the index.
-      // (Subtract 2 because we're going to add two operands after the index.)
-      int maxAllowedRank = kMaxPrintfOperands - printfOperands.size() - 2;
-
-      os << "idx (";
-      const auto &index = indices[i];
-      for (size_t dim = 0; dim < index.size(); dim++) {
-        if (dim != 0) {
-          os << ", ";
-        }
-        if (dim == maxAllowedRank) {
-          os << "... (truncated)";
-          break;
-        }
-        os << getFormatSubstr(index[dim], /*width=*/dimWidths[dim]);
-        printfOperands.push_back(index[dim]);
-      }
-      os << ")";
-
-      os << "%s";
-      printfOperands.push_back(prefixStr);
-
-      if (numOperands > 1) {
-        os << "(operand " << operand << ") ";
-      }
-
-      auto elem = elems[i];
-      os << getFormatSubstr(elem);
-      printfOperands.push_back(elem);
-
-      // It's the same format string each iteration, but it's a lot easier if we
-      // construct the format string at the same time as we populate
-      // printfOperands.  But we don't want to create BLOCK_SIZE duplicate
-      // strings, so we cache the Value.
-      if (i == 0) {
-        formatStrValue = llPrintf(formatStr, printfOperands, rewriter);
-      } else {
-        llPrintf(formatStrValue, printfOperands, rewriter);
-      }
-    }
+    // Ask for 16B alignment on global_smem because that's the largest we should
+    // ever need (4xi32).
+    auto arrayTy = LLVM::LLVMArrayType::get(elemTy, 0);
+    auto global = b.create<LLVM::GlobalOp>(
+        loc, arrayTy, /*isConstant=*/false, LLVM::Linkage::External,
+        "global_smem", /*value=*/Attribute(), /*alignment=*/16,
+        // Add ROCm support.
+        static_cast<unsigned>(NVVM::NVVMMemorySpace::kSharedMemorySpace));
   }
 
-  std::string getFormatSubstr(Value value,
-                              std::optional<int> width = std::nullopt) const {
-    std::string prefix = "%";
-    if (width.has_value()) {
-      prefix += std::to_string(*width);
-    }
-
-    Type type = value.getType();
-    if (type.isa<LLVM::LLVMPointerType>()) {
-      return prefix + "p";
-    } else if (type.isBF16() || type.isF16() || type.isF32() || type.isF64()) {
-      return prefix + "f";
-    } else if (type.isSignedInteger()) {
-      if (type.getIntOrFloatBitWidth() == 64)
-        return prefix + "lli";
-      else
-        return prefix + "i";
-    } else if (type.isUnsignedInteger() || type.isSignlessInteger()) {
-      if (type.getIntOrFloatBitWidth() == 64)
-        return prefix + "llu";
-      else
-        return prefix + "u";
-    }
-    assert(false && "not supported type");
-    return "";
-  }
-
-  // declare vprintf(i8*, i8*) as external function
-  static LLVM::LLVMFuncOp
-  getVprintfDeclaration(ConversionPatternRewriter &rewriter) {
-    auto moduleOp =
-        rewriter.getBlock()->getParent()->getParentOfType<ModuleOp>();
-    StringRef funcName("vprintf");
-    Operation *funcOp = moduleOp.lookupSymbol(funcName);
-    if (funcOp)
-      return cast<LLVM::LLVMFuncOp>(*funcOp);
-
-    auto *context = rewriter.getContext();
-
-    SmallVector<Type> argsType{ptr_ty(context), ptr_ty(context)};
-    auto funcType = LLVM::LLVMFunctionType::get(i32_ty, argsType);
-
-    ConversionPatternRewriter::InsertionGuard guard(rewriter);
-    rewriter.setInsertionPointToStart(moduleOp.getBody());
-
-    return rewriter.create<LLVM::LLVMFuncOp>(UnknownLoc::get(context), funcName,
-                                             funcType);
-  }
-
-  // extend integer to int32, extend float to float64
-  // this comes from vprintf alignment requirements.
-  static std::pair<Type, Value>
-  promoteValue(ConversionPatternRewriter &rewriter, Value value) {
-    auto *context = rewriter.getContext();
-    auto type = value.getType();
-    Value newOp = value;
-    Type newType = type;
-    auto loc = UnknownLoc::get(context);
-
-    bool bUnsigned = type.isUnsignedInteger();
-    if (type.isIntOrIndex() && type.getIntOrFloatBitWidth() < 32) {
-      if (bUnsigned) {
-        newType = ui32_ty;
-        newOp = zext(newType, value);
-      } else {
-        newType = i32_ty;
-        newOp = sext(newType, value);
-      }
-    } else if (type.isBF16() || type.isF16() || type.isF32()) {
-      newType = f64_ty;
-      newOp = fpext(newType, value);
-    }
-
-    return {newType, newOp};
-  }
-
-  // Returns a Value for the format string, which you can reuse.
-  static Value llPrintf(StringRef msg, ValueRange args,
-                        ConversionPatternRewriter &rewriter) {
-    assert(!msg.empty() && "printf with empty string not supported");
-    llvm::SmallString<64> msgNewline(msg);
-    msgNewline.push_back('\n');
-    msgNewline.push_back('\0');
-    Value msgValue =
-        LLVM::addStringToModule(UnknownLoc::get(rewriter.getContext()),
-                                rewriter, "printfFormat_", msgNewline);
-    llPrintf(msgValue, args, rewriter);
-    return msgValue;
-  }
-
-  static void llPrintf(Value msg, ValueRange args,
-                       ConversionPatternRewriter &rewriter) {
-    auto *ctx = rewriter.getContext();
-    Type ptr = ptr_ty(ctx);
-    auto moduleOp =
-        rewriter.getBlock()->getParent()->getParentOfType<ModuleOp>();
-    auto funcOp = getVprintfDeclaration(rewriter);
-    auto loc = UnknownLoc::get(ctx);
-
-    Value one = i32_val(1);
-    Value zero = i32_val(0);
-
-    Value bufferPtr = null(ptr);
-
-    SmallVector<Value, 16> newArgs;
-    if (args.size() >= 1) {
-      SmallVector<Type> argTypes;
-      for (auto arg : args) {
-        Type newType;
-        Value newArg;
-        std::tie(newType, newArg) = promoteValue(rewriter, arg);
-        argTypes.push_back(newType);
-        newArgs.push_back(newArg);
-      }
-
-      Type structTy = LLVM::LLVMStructType::getLiteral(ctx, argTypes);
-      auto allocated =
-          rewriter.create<LLVM::AllocaOp>(loc, ptr_ty(ctx), structTy, one,
-                                          /*alignment=*/0);
-
-      for (const auto &entry : llvm::enumerate(newArgs)) {
-        auto index = i32_val(entry.index());
-        auto fieldPtr =
-            gep(ptr_ty(ctx), structTy, allocated, ArrayRef<Value>{zero, index});
-        store(entry.value(), fieldPtr);
-      }
-      bufferPtr = bitcast(allocated, ptr);
-    }
-
-    SmallVector<Value> operands{msg, bufferPtr};
-    call(funcOp, operands);
+  static Value promoteOperand(OpBuilder &builder, Location loc, Value operand,
+                              Type promotedType) {
+    Type tensorPromotedType =
+        operand.getType().cast<RankedTensorType>().cloneWith(std::nullopt,
+                                                             promotedType);
+    return builder.create<triton::FpToFpOp>(loc, tensorPromotedType, operand);
   }
 };
 
-struct AssertOpConversion : public ConvertOpToLLVMPattern<triton::AssertOp> {
-  using ConvertOpToLLVMPattern<triton::AssertOp>::ConvertOpToLLVMPattern;
+} // anonymous namespace
 
-  LogicalResult
-  matchAndRewrite(triton::AssertOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    auto loc = op.getLoc();
-    auto ctx = rewriter.getContext();
-    auto typeConverter = getTypeConverter();
-    auto elems = unpackLLElements(loc, adaptor.getCondition(), rewriter);
-    auto elemTy = elems[0].getType();
-    Value condition = int_val(elemTy.getIntOrFloatBitWidth(), 0);
-    for (auto elem : elems) {
-      if (elemTy.isSignedInteger() || elemTy.isSignlessInteger()) {
-        condition =
-            or_(condition,
-                icmp_eq(elem, rewriter.create<LLVM::ConstantOp>(
-                                  loc, elemTy, rewriter.getZeroAttr(elemTy))));
-      } else {
-        assert(false && "Unsupported type for assert");
-        return failure();
-      }
-    }
-    llAssert(op, condition, adaptor.getMessage(), adaptor.getFile(),
-             adaptor.getFunc(), adaptor.getLine(), rewriter);
-    rewriter.eraseOp(op);
-    return success();
-  }
+namespace mlir {
+namespace triton {
 
-  // op: the op at which the assert is inserted. Unlike printf, we need to
-  // know about the op to split the block.
-  static void llAssert(Operation *op, Value condition, StringRef message,
-                       StringRef file, StringRef func, int line,
-                       ConversionPatternRewriter &rewriter) {
-    ConversionPatternRewriter::InsertionGuard guard(rewriter);
-    auto ctx = rewriter.getContext();
-    auto loc = op->getLoc();
-
-    // #block1
-    // if (condition) {
-    //   #block2
-    //   __assertfail(message);
-    // }
-    // #block3
-    Block *prevBlock = op->getBlock();
-    Block *ifBlock = rewriter.splitBlock(prevBlock, op->getIterator());
-    rewriter.setInsertionPointToStart(ifBlock);
-
-    auto funcOp = getAssertfailDeclaration(rewriter);
-    auto moduleOp =
-        rewriter.getBlock()->getParent()->getParentOfType<ModuleOp>();
-    Value messageString =
-        LLVM::addStringToModule(loc, rewriter, "assertMessage_", message);
-    Value fileString =
-        LLVM::addStringToModule(loc, rewriter, "assertFile_", file);
-    Value funcString =
-        LLVM::addStringToModule(loc, rewriter, "assertFunc_", func);
-    Value lineNumber = i32_val(line);
-    Value charSize = int_val(sizeof(size_t) * 8, sizeof(char));
-
-    SmallVector<Value> operands = {messageString, fileString, lineNumber,
-                                   funcString, charSize};
-    auto ret = call(funcOp, operands);
-
-    // Split a block after the call.
-    Block *thenBlock = rewriter.splitBlock(ifBlock, op->getIterator());
-    rewriter.setInsertionPointToEnd(ifBlock);
-    rewriter.create<cf::BranchOp>(loc, thenBlock);
-    rewriter.setInsertionPointToEnd(prevBlock);
-    rewriter.create<cf::CondBranchOp>(loc, condition, ifBlock, thenBlock);
-  }
-
-  static LLVM::LLVMFuncOp
-  getAssertfailDeclaration(ConversionPatternRewriter &rewriter) {
-    auto moduleOp =
-        rewriter.getBlock()->getParent()->getParentOfType<ModuleOp>();
-    StringRef funcName("__assertfail");
-    Operation *funcOp = moduleOp.lookupSymbol(funcName);
-    if (funcOp)
-      return cast<LLVM::LLVMFuncOp>(*funcOp);
-
-    // void __assert_fail(const char * assertion, const char * file, unsigned
-    // int line, const char * function);
-    auto *ctx = rewriter.getContext();
-    SmallVector<Type> argsType{ptr_ty(ctx), ptr_ty(ctx), i32_ty, ptr_ty(ctx),
-                               rewriter.getIntegerType(sizeof(size_t) * 8)};
-    auto funcType = LLVM::LLVMFunctionType::get(void_ty(ctx), argsType);
-
-    ConversionPatternRewriter::InsertionGuard guard(rewriter);
-    rewriter.setInsertionPointToStart(moduleOp.getBody());
-
-    return rewriter.create<LLVM::LLVMFuncOp>(UnknownLoc::get(ctx), funcName,
-                                             funcType);
-  }
-};
-
-struct MakeRangeOpConversion
-    : public ConvertOpToLLVMPattern<triton::MakeRangeOp> {
-
-  MakeRangeOpConversion(LLVMTypeConverter &converter, PatternBenefit benefit)
-      : ConvertOpToLLVMPattern<triton::MakeRangeOp>(converter, benefit) {}
-
-  LogicalResult
-  matchAndRewrite(triton::MakeRangeOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    Location loc = op->getLoc();
-    auto rankedTy = op.getResult().getType().cast<RankedTensorType>();
-    auto shape = rankedTy.getShape();
-    auto layout = rankedTy.getEncoding();
-
-    auto elemTy = rankedTy.getElementType();
-    assert(elemTy.isInteger(32));
-    Value start = createIndexAttrConstant(rewriter, loc, elemTy, op.getStart());
-    auto idxs = emitIndices(loc, rewriter, layout, rankedTy, true);
-    unsigned elems = idxs.size();
-    SmallVector<Value> retVals(elems);
-    // TODO: slice layout has more elements than expected.
-    // Unexpected behavior for make range, but generally OK when followed by
-    // expand dims + broadcast. very weird behavior otherwise potentially.
-    for (const auto &multiDim : llvm::enumerate(idxs)) {
-      assert(multiDim.value().size() == 1);
-      retVals[multiDim.index()] = add(multiDim.value()[0], start);
-    }
-    auto typeConverter = getTypeConverter();
-    Value result =
-        packLLElements(loc, typeConverter, retVals, rewriter, rankedTy);
-    rewriter.replaceOp(op, result);
-    return success();
-  }
-};
-
-struct GetProgramIdOpConversion
-    : public ConvertOpToLLVMPattern<triton::GetProgramIdOp> {
-  using ConvertOpToLLVMPattern<triton::GetProgramIdOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::GetProgramIdOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    Value programId = llGetPid(op.getAxisAsInt(), op->getLoc(),
-                               op->getParentOfType<ModuleOp>(), rewriter);
-    rewriter.replaceOp(op, programId);
-    return success();
-  }
-};
-
-struct GetNumProgramsOpConversion
-    : public ConvertOpToLLVMPattern<triton::GetNumProgramsOp> {
-  using ConvertOpToLLVMPattern<
-      triton::GetNumProgramsOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::GetNumProgramsOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    // It is not easy to get the compute capability here, so we use numCTAs to
-    // decide the semantic of GetNumProgramsOp. If numCTAs = 1, then
-    // GetNumProgramsOp is converted to "%nctaid", otherwise it is converted to
-    // "%nclusterid".
-    auto moduleOp = op->getParentOfType<ModuleOp>();
-    assert(moduleOp && "Parent ModuleOp not found for GetProgramIdOp");
-    int numCTAs = triton::gpu::TritonGPUDialect::getNumCTAs(moduleOp);
-
-    Location loc = op->getLoc();
-    assert(op.getAxis() < 3);
-    std::string sreg = numCTAs == 1 ? "%nctaid." : "%nclusterid.";
-    sreg.append(1, 'x' + op.getAxis()); // 0 -> 'x', 1 -> 'y', 2 -> 'z'
-
-    Value numPrograms = getSRegValue(rewriter, loc, sreg);
-    rewriter.replaceOp(op, numPrograms);
-    return success();
-  }
-};
-
-// TODO[goostavz]: GetThreadIdOp/GetClusterCTAIdOp is a temporary solution
-// before async dialect is done. These concepts should appear in ttgpu
-// level, and they are planned to be deprecated along with ttgpu.mbarrier_xxx
-// ops.
-struct GetThreadIdOpConversion
-    : public ConvertOpToLLVMPattern<triton::nvidia_gpu::GetThreadIdOp> {
-  using ConvertOpToLLVMPattern<
-      triton::nvidia_gpu::GetThreadIdOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::nvidia_gpu::GetThreadIdOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    rewriter.replaceOp(op, getThreadId(rewriter, op->getLoc()));
-    return success();
-  }
-};
-
-struct GetCanonicalWarpIdConversion
-    : public ConvertOpToLLVMPattern<triton::nvidia_gpu::GetCanonicalWarpId> {
-  using ConvertOpToLLVMPattern<
-      triton::nvidia_gpu::GetCanonicalWarpId>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::nvidia_gpu::GetCanonicalWarpId op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    rewriter.replaceOp(op, GetCanonicalWarpId(rewriter, op->getLoc()));
-    return success();
-  }
-};
-
-struct GetClusterCTAIdOpConversion
-    : public ConvertOpToLLVMPattern<triton::nvidia_gpu::GetClusterCTAIdOp> {
-  using ConvertOpToLLVMPattern<
-      triton::nvidia_gpu::GetClusterCTAIdOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::nvidia_gpu::GetClusterCTAIdOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    rewriter.replaceOp(op, getClusterCTAId(rewriter, op->getLoc()));
-    return success();
-  }
-};
-
-struct AddPtrOpConversion : public ConvertOpToLLVMPattern<triton::AddPtrOp> {
-  using ConvertOpToLLVMPattern<triton::AddPtrOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::AddPtrOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    Location loc = op->getLoc();
-    auto resultTy = op.getType();
-    auto offsetTy = op.getOffset().getType();
-    auto typeConverter = getTypeConverter();
-    auto resultTensorTy = resultTy.dyn_cast<RankedTensorType>();
-    if (resultTensorTy) {
-      unsigned elems = getTotalElemsPerThread(resultTy);
-      Type elemTy = typeConverter->convertType(resultTensorTy.getElementType()
-                                                   .cast<triton::PointerType>()
-                                                   .getPointeeType());
-      Type ptrTy = typeConverter->convertType(resultTensorTy.getElementType());
-      auto ptrs = unpackLLElements(loc, adaptor.getPtr(), rewriter);
-      auto offsets = unpackLLElements(loc, adaptor.getOffset(), rewriter);
-      SmallVector<Value> resultVals(elems);
-      for (unsigned i = 0; i < elems; ++i) {
-        resultVals[i] = gep(ptrTy, elemTy, ptrs[i], offsets[i]);
-      }
-      Value view =
-          packLLElements(loc, typeConverter, resultVals, rewriter, resultTy);
-      rewriter.replaceOp(op, view);
-    } else {
-      assert(resultTy.isa<triton::PointerType>());
-      auto resultPtrTy = typeConverter->convertType(resultTy);
-      auto resultElemTy = typeConverter->convertType(
-          resultTy.cast<triton::PointerType>().getPointeeType());
-      Value result =
-          gep(resultPtrTy, resultElemTy, adaptor.getPtr(), adaptor.getOffset());
-      rewriter.replaceOp(op, result);
-    }
-    return success();
-  }
-};
-
-struct AllocTensorOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::AllocTensorOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::AllocTensorOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::AllocTensorOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    Location loc = op->getLoc();
-    Value smemBase =
-        LLVM::getSharedMemoryBase(loc, rewriter, op.getOperation());
-    auto resultTy = op.getType().dyn_cast<RankedTensorType>();
-    auto elemPtrTy = ptr_ty(rewriter.getContext(), 3);
-    auto typeConverter = getTypeConverter();
-    smemBase = bitcast(smemBase, elemPtrTy);
-    auto sharedLayout = resultTy.getEncoding().cast<SharedEncodingAttr>();
-    auto order = sharedLayout.getOrder();
-    // Workaround for 3D tensors
-    // TODO: we need to modify the pipeline pass to give a proper shared
-    // encoding to 3D tensors
-    SmallVector<unsigned> newOrder;
-    if (resultTy.getShape().size() != order.size()) {
-      for (auto i = 0; i < order.size(); ++i)
-        newOrder.push_back(order[i] + 1);
-      newOrder.push_back(0);
-    } else {
-      newOrder = SmallVector<unsigned>(order.begin(), order.end());
-    }
-
-    auto llvmElemTy = typeConverter->convertType(resultTy.getElementType());
-    auto shapePerCTA = getShapePerCTA(sharedLayout, resultTy.getShape());
-    auto smemObj = SharedMemoryObject(smemBase, llvmElemTy, shapePerCTA,
-                                      newOrder, loc, rewriter);
-    auto retVal = getStructFromSharedMemoryObject(loc, smemObj, rewriter);
-    rewriter.replaceOp(op, retVal);
-    return success();
-  }
-};
-
-struct DeallocTensorOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::DeallocTensorOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::DeallocTensorOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::DeallocTensorOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-struct ExtractSliceOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::ExtractSliceOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::ExtractSliceOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::ExtractSliceOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    // %dst = extract_slice %src[%offsets]
-    Location loc = op->getLoc();
-    auto srcTy = op.getSource().getType().dyn_cast<RankedTensorType>();
-    auto srcLayout = srcTy.getEncoding().dyn_cast<SharedEncodingAttr>();
-    assert(srcLayout && "Unexpected resultLayout in ExtractSliceOpConversion");
-    assert(op.hasUnitStride() &&
-           "Only unit stride supported by ExtractSliceOpConversion");
-
-    auto typeConverter = getTypeConverter();
-    auto llvmElemTy = typeConverter->convertType(srcTy.getElementType());
-
-    // newBase = base + offset
-    // Triton supports either static and dynamic offsets
-    auto smemObj = LLVM::getSharedMemoryObjectFromStruct(
-        loc, adaptor.getSource(), llvmElemTy, rewriter);
-    SmallVector<Value, 4> opOffsetVals;
-    SmallVector<Value, 4> offsetVals;
-    auto mixedOffsets = op.getMixedOffsets();
-    for (auto i = 0, j = 0; i < mixedOffsets.size(); ++i) {
-      if (op.isDynamicOffset(i)) {
-        // adaptor.getOffsets() returns list of variable offsets. the size of
-        // the list may not be the same as mixedOffsets
-        opOffsetVals.emplace_back(adaptor.getOffsets()[j]);
-        ++j;
-      } else
-        opOffsetVals.emplace_back(i32_val(op.getStaticOffset(i)));
-      offsetVals.emplace_back(add(smemObj.offsets[i], opOffsetVals[i]));
-    }
-    // Compute the offset based on the original strides of the shared memory
-    // object
-    auto offset = dot(rewriter, loc, opOffsetVals, smemObj.strides);
-    // newShape = rank_reduce(shape)
-    // Triton only supports static tensor sizes
-    SmallVector<Value, 4> strideVals;
-    for (auto i = 0; i < op.getStaticSizes().size(); ++i) {
-      if (op.getStaticSize(i) == 1) {
-        offsetVals.erase(offsetVals.begin() + i);
-      } else {
-        strideVals.emplace_back(smemObj.strides[i]);
-      }
-    }
-
-    auto elemPtrTy = ptr_ty(rewriter.getContext(), 3);
-    smemObj =
-        SharedMemoryObject(gep(elemPtrTy, llvmElemTy, smemObj.base, offset),
-                           llvmElemTy, strideVals, offsetVals);
-    auto retVal = getStructFromSharedMemoryObject(loc, smemObj, rewriter);
-    rewriter.replaceOp(op, retVal);
-    return success();
-  }
-};
-
-struct AsyncWaitOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::AsyncWaitOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::AsyncWaitOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::AsyncWaitOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    PTXBuilder ptxBuilder;
-    auto &asyncWaitOp = *ptxBuilder.create<>("cp.async.wait_group");
-    auto num = op->getAttrOfType<IntegerAttr>("num").getInt();
-    asyncWaitOp(ptxBuilder.newConstantOperand(num));
-
-    auto ctx = op.getContext();
-    auto loc = op.getLoc();
-    auto voidTy = void_ty(ctx);
-    ptxBuilder.launch(rewriter, loc, voidTy);
-
-    // Safe to remove the op since it doesn't have any return value.
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-struct AsyncCommitGroupOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::AsyncCommitGroupOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::AsyncCommitGroupOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::AsyncCommitGroupOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-
-    PTXBuilder ptxBuilder;
-    ptxBuilder.create<>("cp.async.commit_group")->operator()();
-    ptxBuilder.launch(rewriter, op.getLoc(), void_ty(op.getContext()));
-    // Safe to remove the op since it doesn't have any return value.
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-struct AsyncBulkWaitOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::AsyncBulkWaitOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::AsyncBulkWaitOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::AsyncBulkWaitOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    PTXBuilder ptxBuilder;
-    auto &asyncBulkWaitOp = *ptxBuilder.create<>("cp.async.bulk.wait_group");
-    auto num = op->getAttrOfType<IntegerAttr>("num").getInt();
-    asyncBulkWaitOp(ptxBuilder.newConstantOperand(num));
-
-    auto ctx = op.getContext();
-    auto loc = op.getLoc();
-    auto voidTy = void_ty(ctx);
-    ptxBuilder.launch(rewriter, loc, voidTy);
-
-    // Safe to remove the op since it doesn't have any return value.
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-struct AsyncBulkCommitGroupOpConversion
-    : public ConvertOpToLLVMPattern<triton::gpu::AsyncBulkCommitGroupOp> {
-  using ConvertOpToLLVMPattern<
-      triton::gpu::AsyncBulkCommitGroupOp>::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::gpu::AsyncBulkCommitGroupOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-
-    PTXBuilder ptxBuilder;
-    ptxBuilder.create<>("cp.async.bulk.commit_group")->operator()();
-    ptxBuilder.launch(rewriter, op.getLoc(), void_ty(op.getContext()));
-    // Safe to remove the op since it doesn't have any return value.
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-} // namespace
-
-void mlir::triton::populateTritonGPUToLLVMPatterns(
-    LLVMTypeConverter &typeConverter, RewritePatternSet &patterns, int numWarps,
-    ModuleAxisInfoAnalysis &axisInfoAnalysis, PatternBenefit benefit) {
-  patterns.add<AddPtrOpConversion>(typeConverter, benefit);
-  patterns.add<AllocTensorOpConversion>(typeConverter, benefit);
-  patterns.add<DeallocTensorOpConversion>(typeConverter, benefit);
-  patterns.add<AsyncCommitGroupOpConversion>(typeConverter, benefit);
-  patterns.add<AsyncWaitOpConversion>(typeConverter, benefit);
-  patterns.add<AsyncBulkCommitGroupOpConversion>(typeConverter, benefit);
-  patterns.add<AsyncBulkWaitOpConversion>(typeConverter, benefit);
-  patterns.add<BroadcastOpConversion>(typeConverter, benefit);
-  patterns.add<ExtractSliceOpConversion>(typeConverter, benefit);
-  patterns.add<GetProgramIdOpConversion>(typeConverter, benefit);
-  patterns.add<GetNumProgramsOpConversion>(typeConverter, benefit);
-  patterns.add<GetThreadIdOpConversion>(typeConverter, benefit);
-  patterns.add<GetCanonicalWarpIdConversion>(typeConverter, benefit);
-  patterns.add<GetClusterCTAIdOpConversion>(typeConverter, benefit);
-  patterns.add<MakeRangeOpConversion>(typeConverter, benefit);
-  patterns.add<ReturnOpConversion>(typeConverter, benefit);
-  patterns.add<PrintOpConversion>(typeConverter, benefit);
-  patterns.add<AssertOpConversion>(typeConverter, benefit);
+std::unique_ptr<OperationPass<ModuleOp>> createConvertTritonGPUToLLVMPass() {
+  return std::make_unique<ConvertTritonGPUToLLVM>();
 }
+std::unique_ptr<OperationPass<ModuleOp>> createConvertTritonGPUToLLVMPass(
+    int32_t computeCapability, Target target,
+    mlir::triton::gpu::TMAMetadataTy *tmaMetadata) {
+  return std::make_unique<ConvertTritonGPUToLLVM>(computeCapability, target,
+                                                  tmaMetadata);
+}
+
+} // namespace triton
+} // namespace mlir
