+++ lib/Conversion/TritonGPUToLLVM/Utility.h
@@ -1221,6 +1221,23 @@ static Value packLLElements(Location loc,
   return llvmStruct;
 }
 
+static Value llGetPid(int axis, Location loc, ModuleOp moduleOp,
+                      ConversionPatternRewriter &rewriter) {
+  assert(axis >= 0);
+  assert(axis < 3);
+  assert(moduleOp);
+
+  // It is not easy to get the compute capability here, so we use numCTAs to
+  // decide the semantic of GetProgramIdOp. If numCTAs = 1, then
+  // GetProgramIdOp is converted to "%ctaid", otherwise it is converted to
+  // "%clusterid".
+  int numCTAs = triton::gpu::TritonGPUDialect::getNumCTAs(moduleOp);
+
+  std::string sreg = numCTAs == 1 ? "%ctaid." : "%clusterid.";
+  sreg.append(1, 'x' + axis); // 0 -> 'x', 1 -> 'y', 2 -> 'z'
+  return LLVM::getSRegValue(rewriter, loc, sreg);
+}
+
 } // namespace mlir
 
 #endif
